requirements.txt
Django==2.1.2
python-memcached==1.59